Various fixes.
- Fix Jacobians producing NaN's
- Support for Ubuntu 20.04
- Other small bugs

---

Jacobian was sometimes NaN when using ceres autodiff.

It looks like Eigen fails to properly extract the rotation from an
Eigen::Transformation (it does SVD which results in NaNs in some
situations. Although I haven't been able to reproduce it manually,
outside ceres.). Explicitly extracting the rotation matrix and
converting it to a Quaternion seems to work though.

Here is an inital pose that resulted in many failed Jacobian entries for
the atrium dataset:

initial_pose:
  pose:
    rotation:
      x: -0.702945
      y: 0.0738337
      z: -0.10968
      w: 0.698847
    translation:
      z: 1.59201
      y: -9.797739999999999
      x: -0.490413
  motion:
    rotation:
      y: 0.103278
      z: -0.247824
      x: 0.0936916
    velocity:
      z: 0.7058680000000001
      x: -0.632722
      y: 0.143362
  stamp:
    sec: 0
    nsec: 123000000